refactor: rename setup/random folder to orderPriority
The folder now holds both ordering (randomization) and prioritization, so
"random" was misleading:
- setup/random -> setup/orderPriority
- RandomError -> OrderPriorityError (it renders both random_group and
  priority_groups instruction errors)
- update imports in SetupPanel and PrioritySetup
