+++ src/components/design/setup/SetupPanel/index.jsx
@@ -31,8 +31,8 @@ import Theming from "../Theming";
 import { ManageLanguages } from "~/pages/manage/ManageTranslations";
 import { useTheme } from "@emotion/react";
 import { questionIconByType } from "~/components/Questions/utils";
-import OrderSetup from "../random/OrderSetup";
-import PrioritySetup from "../random/PrioritySetup";
+import OrderSetup from "../orderPriority/OrderSetup";
+import PrioritySetup from "../orderPriority/PrioritySetup";
 import ScqDefaultValue from "../ScqDefaultValue";
 import DisabledToggle from "../Disabled";
 import MoveGroup from "../MoveGroup";
